Changed setting/settings API.
Documents now only store read in settings, not the defaults. Setting lookup in document checks local and then the defaults.

Both readers and writers can now be intiialised with settings which will
overwrite those associated with the document source. In the case of the
reader the setiings are copied into the document entity. In the case of
writers, they are not and writer handlers should therefore check the
property against the writer, not the document - it will check writer
specific, then document (which in turn will check default).